refactor specs: only define the anon object once
